// projects/shell/src/app/shared/components/table-data/models/table-fields.models.ts
import { MenuActions } from './menu-button-action.models';

export enum TableColorPills {
  RED = '--red',
  GREEN = '--green',
  GRAY = '--gray',
  ORANGE = '--orange',
}

export enum FieldType {
  STRING = 'STRING',
  DATE = 'DATE',
  IMG = 'IMG',
  PILLS = 'PILLS',
  NUMBER = 'NUMBER',
  RADIO = 'RADIO',
  CHECK = 'CHECK',
  MENUACTIONS = 'MENU-ACTIONS',
  DYNAMICMENUACTIONS = 'DYNAMIC MENUACTIONS',
  EXPANDED = 'EXPANDED',
  EXTRA = 'EXTRA',
  EXPANDEDDETAILS = 'EXPANDEDDETAILS',
}

export enum FormatType {
  'YYYY/MM/DD' = 'yyyy/MM/dd',
  'MM/DD/YYYY' = 'MM/yyyy/DD',
  NUMBER = '1.2-2',
}

export interface TableFields {
  name: string /**string name */;
  icon?: string;
  color?: TableColorPills;
  label?: string;
  type: FieldType;
  format?: FormatType | string;
  sorted?: boolean;
  filter?: boolean;
  filterDropDownValues?: {
    label: string;
    value: string | boolean;
  }[];
  widthPercentage?: number;
  rowMenuActionItems?: RowMenuActionItems[];
  value?: any;
}

export interface TableActionsButtons {
  icon: string;
  label: string;
  action: MenuActions;
  disabled?: boolean;
  value?: any;
}

export interface RowMenuActionItems {
  action: RowMenuActions;
}

export enum RowMenuActions {
  VIEWDETAILS = 'View details',
  ASSIGNROLE = 'Assign Role',
  ACTIVATE = 'Activate',
  DEACTIVATE = 'Deactivate',
  CHANGEROLE = 'Change role',
  DELETE = 'Delete',
  PRINT = 'Print',
  PENDINGPRINTING = 'Pending Printing',
  PRINTED = 'Printed',
  COLLECT = 'Collect',
  CANCELED = 'Canceled',
  REJECTED = 'Rejected',
}
export interface TableData {
  currentPageNumber?: number;
  currentPageSize?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
  items: any[];
  itemsCount?: number;
  pageCount?: number;
  lookups?: Lookups;
}

export interface Lookups {
  [key: string]: string[];
}

// Generated by https://quicktype.io

export interface TableFilterObject {
  type?: FieldType;
  key: string;
  value: any;
  display: string;
}